Extracting app setup specific to linux to a different file
